rag/src/requirements.txt
flask~=3.1.0
fastapi~=0.115.12
uvicorn[standard]~=0.34.0
gunicorn
pyyaml~=6.0.2
pyinstaller

pydantic~=2.10.6
schemas~=0.7.1

ollama
openai
google.generativeai
llama-cpp-python

qdrant-client
sentence-transformers
stanza
llama_index
llama_index.embeddings.huggingface
tf-keras
selenium
langdetect
